move create app inside server
